basic authorization

+++ app/Http/Controllers/IdeaController.php
@@ -17,6 +17,8 @@ class IdeaController extends Controller
             'content' => 'required|min:3|max:255',
         ]);
 
+        $validated['user_id'] = auth()->id();
+
         Idea::create($validated);
 
         return redirect()
@@ -25,7 +27,10 @@ class IdeaController extends Controller
     }
     public function destroy(Idea $idea)
     {
-        // Idea::where('id', $id)->firstOrFail()->delete();
+        if(auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
+
         $idea->delete();
 
         return redirect()
@@ -34,12 +39,20 @@ class IdeaController extends Controller
     }
     public function edit(Idea $idea)
     {
+        if(auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
+        
         $editing = true;
         return view('ideas.show', compact('idea', 'editing'));
     }
 
     public function update(Idea $idea)
     {
+        if(auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
+
         $validated = request()->validate([
             'content' => 'required|min:3|max:255',
         ]);
